requirements.txt
pytest==8.3.2
pyyaml==6.0.2
colorama==0.4.6
